economic_calendar_normalizer/requirements.txt
python-dateutil==2.9.0.post0
tzdata==2025.2